Created PSoC 5 Test application and fixed bugs
Fixed some minor bugs in the application code that was preventing the
compiling of the SPIM version of the code.  Also got rid of some
nuisance warnings in util.c

+++ E2ForLife_TerminalDev/PSoC Pioneer + W5500.cydsn/E2ForLife_W5500_v1_0/API/util.c
@@ -136,11 +136,11 @@ cystatus `$INSTANCE_NAME`_ParseMAC(const char *macString, uint8 *mac)
 	index = 0;
 	for(digit = 0;(digit<6) && (result == CYRET_SUCCESS)&&(macString[index] != 0);++digit) {
 		// process the first nibble
-		if (isxdigit(macString[index]) ) {
+		if (isxdigit((int)macString[index]) ) {
 			mac[digit] = _HEX2BIN(macString[index]);
 			++index;
 			mac[digit] <<= 4;
-			if (isxdigit(macString[index])) {
+			if (isxdigit((int)macString[index])) {
 				mac[digit] += _HEX2BIN(macString[index]);
 				++index;
 				/*
